Synchronization: The capuletes and the montagues
In this simulation, a plaza is a critical section that does not permit
both montagues and capuletes present at the same time. In this change, the
server enforces the mutual exclusion by granting appropriate access to the
plaza. The processes forked by the client represent members of the
montagues and capuletes.
NOTE: The IP address of the server and client are not user-specified. The
programs will have to be run on the same machine with the current source
code version. This can be tweaked in the future by specifying a particular
address of interest.